separada responsabilidade de Menu
